# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "lens-analytics-analytics API",
    "description": "Unified Stellar price API aggregating SDEX + AMM prices.",
    "version": "0.1.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/spreads/{asset}": {
      "get": {
        "operationId": "getSpreads",
        "summary": "Get the tightest per-venue bid/ask spread for an asset",
        "description": "Reports, for every (venue, pair) that quoted `asset` inside the window, the observed low, high and spread in basis points, tightest first. The spread is the dispersion of recent prints, not a resting order book's top-of-book gap.",
        "security": [],
        "parameters": [
          {
            "name": "asset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "XLM"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "15m",
                "1h",
                "24h"
              ],
              "default": "5m"
            }
          },
          {
            "$ref": "#/components/parameters/NetworkQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Per-venue spreads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpreadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid window or network"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Spread aggregation failed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "name": "X-PAYMENT",
        "in": "header",
        "description": "Base64 encoded JSON payment proof for x402 micropayments."
      },
      "adminAuth": {
        "type": "apiKey",
        "name": "X-Admin-Key",
        "in": "header"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Client API key issued through the admin API, sent as `Authorization: Bearer <key>`."
      }
    },
    "parameters": {
      "NetworkQuery": {
        "name": "network",
        "in": "query",
        "required": false,
        "description": "Stellar network to answer for — `testnet` or `mainnet`. Defaults to this instance's configured `STELLAR_NETWORK`. An unrecognised value is rejected with 400. DB-backed reads are currently served from whichever network this instance indexes, even when another is requested.",
        "schema": {
          "type": "string",
          "enum": [
            "testnet",
            "mainnet"
          ],
          "default": "testnet"
        }
      }
    },
    "responses": {
...
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait before retrying"
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Remaining requests in current window"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": {
                  "type": "integer"
                },
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "retryAfter": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "SpreadVenue": {
        "type": "object",
        "required": [
          "venue",
          "pairKey",
          "bid",
          "ask",
          "mid",
          "spreadBps",
          "observations",
          "lastSeen"
        ],
        "additionalProperties": false,
        "properties": {
          "venue": {
            "type": "string"
          },
          "pairKey": {
            "type": "string"
          },
          "bid": {
            "type": "number"
          },
          "ask": {
            "type": "number"
          },
          "mid": {
            "type": "number"
          },
          "spreadBps": {
            "type": "number"
          },
          "observations": {
            "type": "integer"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SpreadResponse": {
        "type": "object",
        "required": [
          "asset",
          "network",
          "window",
          "startTime",
          "endTime",
          "venues",
          "tightest"
        ],
        "additionalProperties": false,
        "properties": {
          "asset": {
            "type": "string"
          },
          "network": {
            "type": "string",
            "enum": [
              "testnet",
              "mainnet"
            ]
          },
          "window": {
            "type": "string",
            "enum": [
              "5m",
              "15m",
              "1h",
              "24h"
            ]
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpreadVenue"
            }
          },
          "tightest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SpreadVenue"
              }
            ]
          }
        }
      },
...
    }
  }
}